requirements.txt
flask
ultralytics
transformers
torch
pillow
pytesseract
opencv-python
geopy